게시글 삭제시 권한확인 (로그인 안하면 401, 자기글 아닌데 삭제시 403, 자기글 삭제 200)

+++ src/main/java/com/example/prj1be20231109/controller/BoardController.java
@@ -54,7 +54,16 @@ public class BoardController {
     }
 
     @DeleteMapping("remove/{id}")
-    public ResponseEntity remove(@PathVariable Integer id) {
+    public ResponseEntity remove(@PathVariable Integer id,
+                                 @SessionAttribute (value = "login", required = false) Member login) {
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401오류
+        }
+
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403오류
+        }
+
         if (service.remove(id)) {
             return ResponseEntity.ok().build();
         } else {
